Move installer directory helper into paths utils
