+++ pkg/fleet/installer/paths/installer_paths_utils.go
@@ -12,6 +12,34 @@ import (
 	"os"
 )
 
+// EnsureInstallerDirectories creates the installer data, packages, configs, tmp,
+// and run directories if they do not exist.
+func EnsureInstallerDirectories() error {
+	err := EnsureInstallerDataDir()
+	if err != nil {
+		return fmt.Errorf("could not ensure installer data directory permissions: %w", err)
+	}
+
+	err = os.MkdirAll(PackagesPath, 0755)
+	if err != nil {
+		return fmt.Errorf("error creating packages directory: %w", err)
+	}
+	err = os.MkdirAll(ConfigsPath, 0755)
+	if err != nil {
+		return fmt.Errorf("error creating configs directory: %w", err)
+	}
+	err = os.MkdirAll(RootTmpDir, 0755)
+	if err != nil {
+		return fmt.Errorf("error creating tmp directory: %w", err)
+	}
+	err = os.MkdirAll(RunPath, 0755)
+	if err != nil {
+		return fmt.Errorf("error creating run directory: %w", err)
+	}
+
+	return nil
+}
+
 // CopyFile copies a file from src to dst
 func CopyFile(src, dst string) error {
 	// Open the source file
